+++ DESCRIPTION
@@ -37,28 +37,32 @@ Imports:
   Rsymphony,
   readr,
   tidyverse,
-  reshape2
+  dplyr,
+  reshape2,
+  Rcpp
 Suggests: 
   knitr,
-  dplyr,
   gurobi (>= 9.0),
   roxygen2,
   Rglpk,
   rmarkdown,
   testthat,
   Rcpp
+LinkingTo: 
+  Rcpp, 
+  RcppArmadillo, 
+  BH
 Encoding: UTF-8
 LazyData: true
 SystemRequirements: C++11
 License: GPL-3
-LinkingTo: Rcpp, RcppArmadillo, BH
 RoxygenNote: 7.1.1
 URL: https://prioriactions.github.io/prioriactions, 
   https://github.com/prioriactions/prioriactions
 BugReports: https://github.com/prioriactions/prioriactions/issues
 VignetteBuilder: knitr
 Collate: 
-    'MAMP_model.R'
+    'RcppExports.R'
     'internal.R'
     'conservation_problem.R'
     'createtxt.R'
@@ -71,4 +75,5 @@ Collate:
     'show.R'
     'solution.R'
     'solve.R'
+    'utils-pipe.R'
     'zzz.R'
