+++ NAMESPACE
@@ -8,6 +8,7 @@ S3method(print,Solution)
 S3method(show,ConservationProblem)
 S3method(show,OptimizationProblem)
 S3method(show,Solution)
+export("%>%")
 export()
 export(ConservationProblem)
 export(OptimizationProblem)
@@ -16,4 +17,6 @@ export(min_costs)
 export(problem)
 export(solve)
 import(Matrix)
+import(Rcpp)
+importFrom(magrittr,"%>%")
 useDynLib(prioriactions, .registration = TRUE)
