+++ R/RcppExports.R
@@ -1,0 +1,39 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+rcpp_new_optimization_problem <- function(nrow = 1000000L, ncol = 1000000L, ncell = 100000L) {
+    .Call(`_prioriactions_rcpp_new_optimization_problem`, nrow, ncol, ncell)
+}
+
+rcpp_optimization_problem_as_list <- function(x) {
+    .Call(`_prioriactions_rcpp_optimization_problem_as_list`, x)
+}
+
+rcpp_get_optimization_problem_ncol <- function(x) {
+    .Call(`_prioriactions_rcpp_get_optimization_problem_ncol`, x)
+}
+
+rcpp_get_optimization_problem_nrow <- function(x) {
+    .Call(`_prioriactions_rcpp_get_optimization_problem_nrow`, x)
+}
+
+rcpp_get_optimization_problem_ncell <- function(x) {
+    .Call(`_prioriactions_rcpp_get_optimization_problem_ncell`, x)
+}
+
+rcpp_get_optimization_problem_A <- function(x) {
+    .Call(`_prioriactions_rcpp_get_optimization_problem_A`, x)
+}
+
+rcpp_constraint_benefit <- function(x, pu_data, features_data, dist_features_data, threats_data, dist_threats_data, sensitivity_data, curve, segments) {
+    .Call(`_prioriactions_rcpp_constraint_benefit`, x, pu_data, features_data, dist_features_data, threats_data, dist_threats_data, sensitivity_data, curve, segments)
+}
+
+rcpp_min_set <- function(x, features_data, pu_data, bound_data, dist_features_data, dist_threats_data, sensitivity_data, threats_data, settings_Data) {
+    .Call(`_prioriactions_rcpp_min_set`, x, features_data, pu_data, bound_data, dist_features_data, dist_threats_data, sensitivity_data, threats_data, settings_Data)
+}
+
+rcpp_objective_min_set <- function(x, pu_data, threats_data, dist_threats_data, boundary_data, blm) {
+    .Call(`_prioriactions_rcpp_objective_min_set`, x, pu_data, threats_data, dist_threats_data, boundary_data, blm)
+}
+
