+++ R/conservation_problem.R
@@ -24,7 +24,7 @@ NULL
 #'   object stored in the \code{data} field with the corresponding \code{name}.
 #'   The argument \code{name} indicates the name of arguments of the
 #'   \code{problem} function ("pu", "features", "rij", "threats", "sensitivity"
-#'   or "bound").}
+#'   or "boundary").}
 #'
 #'   \item{$getFeatureAmount( )}{return \code{integer} number of features.}
 #'
@@ -53,18 +53,16 @@ NULL
 #' ## Examples of how to use the methods of a ConservationProblem class object.
 #'
 #' ## Load data
-#' data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensitivity_data, example_bound_data)
+#' data(example_pu_data, example_features_data, example_rij_data, example_threats_data, example_sensitivity_data, example_boundary_data)
 #'
 #' ## Create data instance
 #' problem_data <- problem(
-#'   pu = example_pu_data, features = example_features_data, rij = example_rij_data,
-#'   threats = example_threats_data, sensitivity = example_sensibility_data,
-#'   bound = example_bound_data
+#'   pu = example_pu_data, features = example_features_data, dist_features = example_dist_features__data,
+#'   threats = example_threats_data, dist_therats = example_dist_threats_data, sensitivity = example_sensibility_data,
+#'   boundary = example_boundary_data
 #' )
 #'
 #' ## Use class methods
-#' problem_data$getActionsAmount()
-#'
 #' problem_data$getData("features")
 #'
 #' problem_data$getFeatureAmount()
@@ -157,29 +155,29 @@ ConservationProblem <- pproto(
   },
   getThreatNames = function(self) {
     if (inherits(self$data$threats, "data.frame")) {
-      return(as.character(sort(unique(self$data$threats$threats))))
+      return(as.character(self$data$threats$name))
     } else {
       stop("threats data is of an unrecognized class")
     }
   },
   getThreatsAmount = function(self) {
     if (inherits(self$data$threats, "data.frame")) {
-      return(length(unique(self$data$threats$threats)))
+      return(nrow(self$data$threats))
     } else {
       stop("threats data is of an unrecognized class")
     }
   },
   getThreatCosts = function(self) {
-    if (inherits(self$data$threats, "data.frame")) {
-      m <- as.vector(self$data$threats$cost)
+    if (inherits(self$data$dist_threats, "data.frame")) {
+      m <- as.vector(self$data$dist_threats$cost)
     } else {
       stop("cost is of unknown class")
     }
     return(m)
   },
   getActionsAmount = function(self) {
-    if (inherits(self$data$threats, "data.frame")) {
-      return(nrow(self$data$threats))
+    if (inherits(self$data$dist_threats, "data.frame")) {
+      return(nrow(self$data$dist_threats))
     } else {
       stop("threats data is of an unrecognized class")
     }
