+++ inst/extdata/input/dist_threats.dat
@@ -1,0 +1,563 @@
+pu,threats,amount,cost,status
+1,2,1,3,0
+1,4,1,3,0
+2,2,1,4,0
+2,3,1,4,0
+2,4,1,4,0
+3,2,1,3,0
+3,4,1,3,0
+4,2,1,3,0
+4,4,1,3,0
+5,2,1,3,0
+5,4,1,3,0
+6,2,1,3,0
+6,4,1,3,0
+7,2,1,3,0
+7,4,1,3,0
+8,2,1,3,0
+8,4,1,3,0
+9,2,1,3,0
+9,4,1,3,0
+10,2,1,3,0
+10,4,1,3,0
+11,2,1,3,0
+11,4,1,3,0
+12,2,1,3,0
+12,4,1,3,0
+13,2,1,3,0
+13,4,1,3,0
+14,2,1,3,0
+14,4,1,3,0
+15,2,1,3,0
+15,4,1,3,0
+16,2,1,3,0
+16,4,1,3,0
+17,2,1,3,0
+17,4,1,3,0
+18,2,1,3,0
+18,4,1,3,0
+19,2,1,3,0
+19,4,1,3,0
+20,2,1,3,0
+20,4,1,3,0
+21,2,1,3,0
+21,4,1,3,0
+22,2,1,3,0
+22,4,1,3,0
+23,2,1,3,0
+23,4,1,3,0
+24,2,1,3,0
+24,4,1,3,0
+25,2,1,3,0
+25,4,1,3,0
+26,2,1,3,0
+26,4,1,3,0
+27,2,1,3,0
+27,4,1,3,0
+28,2,1,3,0
+28,4,1,3,0
+29,2,1,3,0
+29,4,1,3,0
+30,2,1,3,0
+30,4,1,3,0
+31,2,1,3,0
+31,4,1,3,0
+32,2,1,3,0
+32,4,1,3,0
+33,2,1,3,0
+33,4,1,3,0
+34,2,1,3,0
+34,4,1,3,0
+35,2,1,3,0
+35,4,1,3,0
+36,2,1,3,0
+36,4,1,3,0
+37,2,1,3,0
+37,4,1,3,0
+38,2,1,3,0
+38,4,1,3,0
+39,2,1,3,0
+39,4,1,3,0
+40,2,1,3,0
+40,4,1,3,0
+41,2,1,3,0
+41,4,1,3,0
+42,2,1,3,0
+42,4,1,3,0
+43,2,1,3,0
+43,4,1,3,0
+44,2,1,3,0
+44,4,1,3,0
+45,2,1,3,0
+45,4,1,3,0
+46,2,1,3,0
+46,4,1,3,0
+47,2,1,3,0
+47,4,1,3,0
+48,2,1,3,0
+48,4,1,3,0
+49,2,1,3,0
+49,4,1,3,0
+50,2,1,3,0
+50,4,1,3,0
+51,2,1,3,0
+51,4,1,3,0
+52,2,1,4,0
+52,3,1,4,0
+52,4,1,4,0
+53,2,1,3,0
+53,4,1,3,0
+54,2,1,3,0
+54,4,1,3,0
+55,2,1,3,0
+55,4,1,3,0
+56,2,1,3,0
+56,4,1,3,0
+57,2,1,3,0
+57,4,1,3,0
+58,2,1,3,0
+58,4,1,3,0
+59,1,1,5,0
+59,2,1,5,0
+59,3,1,5,0
+59,4,1,5,0
+60,2,1,3,0
+60,4,1,3,0
+61,2,1,3,0
+61,4,1,3,0
+62,2,1,3,0
+62,4,1,3,0
+63,1,1,5,0
+63,2,1,5,0
+63,3,1,5,0
+63,4,1,5,0
+64,2,1,3,0
+64,4,1,3,0
+65,2,1,3,0
+65,4,1,3,0
+66,2,1,3,0
+66,4,1,3,0
+67,2,1,3,0
+67,4,1,3,0
+68,2,1,3,0
+68,4,1,3,0
+69,2,1,3,0
+69,4,1,3,0
+70,2,1,3,0
+70,4,1,3,0
+71,2,1,3,0
+71,4,1,3,0
+72,2,1,3,0
+72,4,1,3,0
+73,2,1,3,0
+73,4,1,3,0
+74,2,1,3,0
+74,4,1,3,0
+75,2,1,3,0
+75,4,1,3,0
+76,2,1,3,0
+76,4,1,3,0
+77,2,1,3,0
+77,4,1,3,0
+78,2,1,3,0
+78,4,1,3,0
+79,2,1,3,0
+79,4,1,3,0
+80,2,1,3,0
+80,4,1,3,0
+81,2,1,3,0
+81,4,1,3,0
+82,2,1,3,0
+82,4,1,3,0
+83,2,1,3,0
+83,4,1,3,0
+84,2,1,3,0
+84,4,1,3,0
+85,2,1,3,0
+85,4,1,3,0
+86,2,1,3,0
+86,4,1,3,0
+87,2,1,3,0
+87,4,1,3,0
+88,2,1,3,0
+88,4,1,3,0
+89,2,1,3,0
+89,4,1,3,0
+90,2,1,3,0
+90,4,1,3,0
+91,2,1,3,0
+91,4,1,3,0
+92,2,1,3,0
+92,4,1,3,0
+93,2,1,3,0
+93,4,1,3,0
+94,2,1,3,0
+94,4,1,3,0
+95,2,1,3,0
+95,4,1,3,0
+96,2,1,4,0
+96,3,1,4,0
+96,4,1,4,0
+97,2,1,4,0
+97,3,1,4,0
+97,4,1,4,0
+98,2,1,3,0
+98,4,1,3,0
+99,2,1,3,0
+99,4,1,3,0
+100,2,1,3,0
+100,4,1,3,0
+101,2,1,3,0
+101,4,1,3,0
+102,2,1,3,0
+102,4,1,3,0
+103,2,1,3,0
+103,4,1,3,0
+104,1,1,5,0
+104,2,1,5,0
+104,3,1,5,0
+104,4,1,5,0
+105,1,1,5,0
+105,2,1,5,0
+105,3,1,5,0
+105,4,1,5,0
+106,2,1,3,0
+106,4,1,3,0
+107,2,1,3,0
+107,4,1,3,0
+108,2,1,3,0
+108,4,1,3,0
+109,1,1,5,0
+109,2,1,5,0
+109,3,1,5,0
+109,4,1,5,0
+110,2,1,3,0
+110,4,1,3,0
+111,2,1,3,0
+111,4,1,3,0
+112,2,1,3,0
+112,4,1,3,0
+113,2,1,3,0
+113,4,1,3,0
+114,2,1,3,0
+114,4,1,3,0
+115,2,1,3,0
+115,4,1,3,0
+116,2,1,3,0
+116,4,1,3,0
+117,2,1,3,0
+117,4,1,3,0
+118,1,1,5,0
+118,2,1,5,0
+118,3,1,5,0
+118,4,1,5,0
+119,2,1,3,0
+119,4,1,3,0
+120,2,1,3,0
+120,4,1,3,0
+121,1,1,5,0
+121,2,1,5,0
+121,3,1,5,0
+121,4,1,5,0
+122,1,1,5,0
+122,2,1,5,0
+122,3,1,5,0
+122,4,1,5,0
+123,1,1,5,0
+123,2,1,5,0
+123,3,1,5,0
+123,4,1,5,0
+124,2,1,3,0
+124,4,1,3,0
+125,2,1,3,0
+125,4,1,3,0
+126,2,1,4,0
+126,3,1,4,0
+126,4,1,4,0
+127,2,1,3,0
+127,4,1,3,0
+128,2,1,3,0
+128,4,1,3,0
+129,1,1,5,0
+129,2,1,5,0
+129,3,1,5,0
+129,4,1,5,0
+130,2,1,3,0
+130,4,1,3,0
+131,2,1,3,0
+131,4,1,3,0
+132,2,1,3,0
+132,4,1,3,0
+133,2,1,3,0
+133,4,1,3,0
+134,1,1,5,0
+134,2,1,5,0
+134,3,1,5,0
+134,4,1,5,0
+135,2,1,3,0
+135,4,1,3,0
+136,1,1,5,0
+136,2,1,5,0
+136,3,1,5,0
+136,4,1,5,0
+137,2,1,3,0
+137,4,1,3,0
+138,2,1,3,0
+138,4,1,3,0
+139,2,1,3,0
+139,4,1,3,0
+140,2,1,3,0
+140,4,1,3,0
+141,1,1,5,0
+141,2,1,5,0
+141,3,1,5,0
+141,4,1,5,0
+142,2,1,4,0
+142,3,1,4,0
+142,4,1,4,0
+143,2,1,3,0
+143,4,1,3,0
+144,2,1,3,0
+144,4,1,3,0
+145,2,1,3,0
+145,4,1,3,0
+146,2,1,4,0
+146,3,1,4,0
+146,4,1,4,0
+147,2,1,3,0
+147,4,1,3,0
+148,2,1,3,0
+148,4,1,3,0
+149,2,1,3,0
+149,4,1,3,0
+150,2,1,3,0
+150,4,1,3,0
+151,2,1,4,0
+151,3,1,4,0
+151,4,1,4,0
+152,2,1,3,0
+152,4,1,3,0
+153,2,1,4,0
+153,3,1,4,0
+153,4,1,4,0
+154,2,1,4,0
+154,3,1,4,0
+154,4,1,4,0
+155,2,1,3,0
+155,4,1,3,0
+156,2,1,3,0
+156,4,1,3,0
+157,2,1,3,0
+157,4,1,3,0
+158,2,1,3,0
+158,4,1,3,0
+159,2,1,3,0
+159,4,1,3,0
+160,2,1,3,0
+160,4,1,3,0
+161,2,1,3,0
+161,4,1,3,0
+162,2,1,4,0
+162,3,1,4,0
+162,4,1,4,0
+163,2,1,4,0
+163,3,1,4,0
+163,4,1,4,0
+164,2,1,3,0
+164,4,1,3,0
+165,2,1,3,0
+165,4,1,3,0
+166,2,1,4,0
+166,3,1,4,0
+166,4,1,4,0
+167,2,1,4,0
+167,3,1,4,0
+167,4,1,4,0
+168,2,1,3,0
+168,4,1,3,0
+169,2,1,4,0
+169,3,1,4,0
+169,4,1,4,0
+170,2,1,4,0
+170,3,1,4,0
+170,4,1,4,0
+171,2,1,4,0
+171,3,1,4,0
+171,4,1,4,0
+172,2,1,3,0
+172,4,1,3,0
+173,2,1,4,0
+173,3,1,4,0
+173,4,1,4,0
+174,2,1,4,0
+174,3,1,4,0
+174,4,1,4,0
+175,2,1,3,0
+175,4,1,3,0
+176,2,1,4,0
+176,3,1,4,0
+176,4,1,4,0
+177,2,1,4,0
+177,3,1,4,0
+177,4,1,4,0
+178,2,1,3,0
+178,4,1,3,0
+179,2,1,4,0
+179,3,1,4,0
+179,4,1,4,0
+180,2,1,3,0
+180,4,1,3,0
+181,2,1,3,0
+181,4,1,3,0
+182,2,1,3,0
+182,4,1,3,0
+183,2,1,4,0
+183,3,1,4,0
+183,4,1,4,0
+184,2,1,3,0
+184,4,1,3,0
+185,1,1,5,0
+185,2,1,5,0
+185,3,1,5,0
+185,4,1,5,0
+186,2,1,4,0
+186,3,1,4,0
+186,4,1,4,0
+187,2,1,3,0
+187,4,1,3,0
+188,2,1,4,0
+188,3,1,4,0
+188,4,1,4,0
+189,2,1,3,0
+189,4,1,3,0
+190,2,1,4,0
+190,3,1,4,0
+190,4,1,4,0
+191,2,1,4,0
+191,3,1,4,0
+191,4,1,4,0
+192,2,1,4,0
+192,3,1,4,0
+192,4,1,4,0
+193,2,1,3,0
+193,4,1,3,0
+194,1,1,5,0
+194,2,1,5,0
+194,3,1,5,0
+194,4,1,5,0
+195,2,1,3,0
+195,4,1,3,0
+196,1,1,5,0
+196,2,1,5,0
+196,3,1,5,0
+196,4,1,5,0
+197,2,1,3,0
+197,4,1,3,0
+198,1,1,5,0
+198,2,1,5,0
+198,3,1,5,0
+198,4,1,5,0
+199,2,1,4,0
+199,3,1,4,0
+199,4,1,4,0
+200,2,1,3,0
+200,4,1,3,0
+201,2,1,3,0
+201,4,1,3,0
+202,2,1,3,0
+202,4,1,3,0
+203,2,1,4,0
+203,3,1,4,0
+203,4,1,4,0
+204,2,1,3,0
+204,4,1,3,0
+205,2,1,3,0
+205,4,1,3,0
+206,2,1,3,0
+206,4,1,3,0
+207,2,1,3,0
+207,4,1,3,0
+208,2,1,3,0
+208,4,1,3,0
+209,2,1,3,0
+209,4,1,3,0
+210,2,1,4,0
+210,3,1,4,0
+210,4,1,4,0
+211,1,1,5,0
+211,2,1,5,0
+211,3,1,5,0
+211,4,1,5,0
+212,1,1,5,0
+212,2,1,5,0
+212,3,1,5,0
+212,4,1,5,0
+213,2,1,4,0
+213,3,1,4,0
+213,4,1,4,0
+214,2,1,3,0
+214,4,1,3,0
+215,2,1,4,0
+215,3,1,4,0
+215,4,1,4,0
+216,2,1,4,0
+216,3,1,4,0
+216,4,1,4,0
+217,2,1,3,0
+217,4,1,3,0
+218,2,1,3,0
+218,4,1,3,0
+219,2,1,3,0
+219,4,1,3,0
+220,2,1,4,0
+220,3,1,4,0
+220,4,1,4,0
+221,2,1,3,0
+221,4,1,3,0
+222,2,1,3,0
+222,4,1,3,0
+223,2,1,3,0
+223,4,1,3,0
+224,2,1,4,0
+224,3,1,4,0
+224,4,1,4,0
+225,2,1,3,0
+225,4,1,3,0
+226,2,1,3,0
+226,4,1,3,0
+227,2,1,4,0
+227,3,1,4,0
+227,4,1,4,0
+228,2,1,3,0
+228,4,1,3,0
+229,2,1,4,0
+229,3,1,4,0
+229,4,1,4,0
+230,2,1,4,0
+230,3,1,4,0
+230,4,1,4,0
+231,2,1,4,0
+231,3,1,4,0
+231,4,1,4,0
+232,2,1,4,0
+232,3,1,4,0
+232,4,1,4,0
+233,2,1,3,0
+233,4,1,3,0
+234,2,1,3,0
+234,4,1,3,0
+235,2,1,4,0
+235,3,1,4,0
+235,4,1,4,0
+236,2,1,3,0
+236,4,1,3,0
+237,2,1,3,0
+237,4,1,3,0
+238,2,1,3,0
+238,4,1,3,0
+239,2,1,3,0
+239,4,1,3,0
+240,1,1,5,0
+240,2,1,5,0
+240,3,1,5,0
+240,4,1,5,0
