--- inst/extdata/input/sensibility.dat
@@ -1,41 +1,0 @@
-species,threats,amount
-1,1,1
-1,2,1
-1,3,1
-1,4,1
-2,1,1
-2,2,1
-2,3,1
-2,4,1
-3,1,1
-3,2,1
-3,3,1
-3,4,1
-4,1,1
-4,2,1
-4,3,1
-4,4,1
-5,1,1
-5,2,1
-5,3,1
-5,4,1
-6,1,1
-6,2,1
-6,3,1
-6,4,1
-7,1,1
-7,2,1
-7,3,1
-7,4,1
-8,1,1
-8,2,1
-8,3,1
-8,4,1
-9,1,1
-9,2,1
-9,3,1
-9,4,1
-10,1,1
-10,2,1
-10,3,1
-10,4,1
