+++ inst/extdata/input/sensitivity.dat
@@ -1,0 +1,41 @@
+species,threats,a,b,c,d
+1,1,0,1,0,1
+1,2,0,1,0,1
+1,3,0,1,0,1
+1,4,0,1,0,1
+2,1,0,1,0,1
+2,2,0,1,0,1
+2,3,0,1,0,1
+2,4,0,1,0,1
+3,1,0,1,0,1
+3,2,0,1,0,1
+3,3,0,1,0,1
+3,4,0,1,0,1
+4,1,0,1,0,1
+4,2,0,1,0,1
+4,3,0,1,0,1
+4,4,0,1,0,1
+5,1,0,1,0,1
+5,2,0,1,0,1
+5,3,0,1,0,1
+5,4,0,1,0,1
+6,1,0,1,0,1
+6,2,0,1,0,1
+6,3,0,1,0,1
+6,4,0,1,0,1
+7,1,0,1,0,1
+7,2,0,1,0,1
+7,3,0,1,0,1
+7,4,0,1,0,1
+8,1,0,1,0,1
+8,2,0,1,0,1
+8,3,0,1,0,1
+8,4,0,1,0,1
+9,1,0,1,0,1
+9,2,0,1,0,1
+9,3,0,1,0,1
+9,4,0,1,0,1
+10,1,0,1,0,1
+10,2,0,1,0,1
+10,3,0,1,0,1
+10,4,0,1,0,1
