+++ inst/extdata/input_big/.~lock.unitCost_Big.csv#
@@ -1,0 +1,1 @@
+,josesalgr,josesalgr-ThinkPad-E570,09.07.2021 19:13,file:///home/josesalgr/.config/libreoffice/4;
