+++ man/problem.Rd
@@ -4,9 +4,9 @@
 \alias{problem}
 \title{Multi-action conservation planning problem}
 \usage{
-problem(pu, features, rij, threats, sensitivity, ...)
+problem(pu, features, dist_features, threats, dist_threats, sensitivity, ...)
 
-\S4method{problem}{data.frame}(pu, features, rij, threats, sensitivity, bound = NULL)
+\S4method{problem}{data.frame}(pu, features, dist_features, threats, dist_threats, sensitivity, boundary = NULL)
 }
 \arguments{
 \item{pu}{Object of class \code{\link{data.frame}} that specifies the planning units
@@ -33,7 +33,7 @@ corresponds to different information about the features. The description of the 
    \item{\code{"name"}}{\code{character} name for each conservation feature. This column is \strong{optional}.}
    }}
 
-\item{rij}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of conservation features across planning units.
+\item{dist_features}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of conservation features across planning units.
 \strong{It is recommended not to include cases where a feature does not occur in a planning unit (avoid rows with
 \code{"amount"} equal to 0). Instead, it is recommended to omit these cases completely from the \code{\link{data.frame}}, for efficiency reasons.}
 The description of the columns is listed below.
@@ -43,7 +43,16 @@ The description of the columns is listed below.
    \item{\code{"amount"}}{\code{integer} indicating if the conservation feature occurring in the planning unit listed on the same row (1), or if it is not occurring (0). This column is \strong{mandatory}.}
    }}
 
-\item{threats}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of threats across planning units
+\item{threats}{Object of class \code{\link{data.frame}} that specifies the threats to consider in
+the reserve design exercise, such as their name and penalty of connectivity (i.e., \emph{blm_actions}). Each row corresponds to a different feature and each column
+corresponds to different information about the threats The description of the columns is listed below.
+\describe{
+   \item{\code{"id"}}{\code{integer} unique identifier for each threat. This column is \strong{mandatory}.}
+   \item{\code{"name"}}{\code{character} name for each threat. This column is \strong{optional}.}
+   \item{\code{"blm_actions"}}{\code{numeric} penalty of connectivity between threats. This column is \strong{optional}.}
+   }}
+
+\item{dist_threats}{Object of class \code{\link{data.frame}} that specifies the spatial distribution of threats across planning units
 and the corresponding cost of applying an action against these threats. \strong{It is recommended not to include cases where a threat does not occur
 in a planning unit (avoid rows with \code{"amount"} equal to 0). Instead, it is recommended to omit these cases completely from the \code{\link{data.frame}}, for efficiency reasons.}
 The description of the columns is listed below.
@@ -67,7 +76,7 @@ The description of the columns is listed below.
    \item{\code{"amount"}}{\code{integer} indicating if the threat againsting the conservation feature listed on the same row (1), or if it is not againsting (0). This column is \strong{mandatory}.}
    }}
 
-\item{bound}{Object of class \code{\link{NULL}} (default) or an object of class \code{\link{data.frame}}.
+\item{boundary}{Object of class \code{\link{NULL}} (default) or an object of class \code{\link{data.frame}}.
 \code{NULL} indicates that planning units' boundaries (i.e., \emph{boundary data}) are not required for the multi-action conservation planning problem.
 \code{data.frame} specifies the boundaries of the planning units, i.e., the spatial relationship between two units, such as the "length" of the shared boundary between them.
 The description of the columns is listed below.
@@ -93,8 +102,8 @@ If you are familiar with using conservation planning software called
 follow the conventions used by the Marxan input files. This means that,
 \code{pu} corresponds to the data in the \strong{Planning Unit File} (typically called \code{"pu.dat"});
 \code{features} corresponds to the data in the \strong{Conservation Feature File} (typically
-called \code{"spec.dat"}); \code{rij} corresponds to the data in the \strong{Planning Unit versus
-Conservation Feature File} (typically called \code{"puvspr.dat"}); and \code{bound} corresponds
+called \code{"spec.dat"}); \code{dist_features} corresponds to the data in the \strong{Planning Unit versus
+Conservation Feature File} (typically called \code{"dist_features.dat"}); and \code{boundary} corresponds
 to the data in the \strong{Boundary Length File} (typically called \code{"bound.dat"}).
 }
 \examples{
@@ -102,41 +111,42 @@ to the data in the \strong{Boundary Length File} (typically called \code{"bound.
 ## using data.frames that have been loaded into R. This example uses input files included
 ## into package.
 
+## Set prioriactions path
+prioriactions_path <- system.file("extdata/input/", package = "prioriactions")
+
 ## Load in planning unit data
-pu_path <- system.file("extdata/input/pu.dat", package = "prioriactions")
-pu_data <- data.table::fread(pu_path, data.table = FALSE)
+pu_data <- data.table::fread(paste0(prioriactions_path,"pu.dat"), data.table = FALSE)
 head(pu_data)
 
 ## Load in feature data
-spec_path <- system.file("extdata/input/spec.dat", package = "prioriactions")
-spec_data <- data.table::fread(spec_path, data.table = FALSE)
+features_data <- data.table::fread(paste0(prioriactions_path,"features.dat"), data.table = FALSE)
 head(spec_data)
 
 ## Load in planning unit vs feature data
-puvspr_path <- system.file("extdata/input/puvspr.dat", package = "prioriactions")
-puvspr_data <- data.table::fread(puvspr_path, data.table = FALSE)
-head(puvspr_data)
+dist_features_data <- data.table::fread(paste0(prioriactions_path,"dist_features.dat"), data.table = FALSE)
+head(dist_features_data)
 
 ## Load in the threats data
-threats_path <- system.file("extdata/input/threats.dat", package = "prioriactions")
-threats_data <- data.table::fread(threats_path, data.table = FALSE)
+threats_data <- data.table::fread(paste0(prioriactions_path,"threats.dat"), data.table = FALSE)
 head(threats_data)
 
+## Load in the threats distribution data
+dist_threats_data <- data.table::fread(paste0(prioriactions_path,"dist_threats.dat"), data.table = FALSE)
+head(dist_threats_data)
+
 ## Load in the sensitivity data
-sensitivity_path <- system.file("extdata/input/sensibility.dat", package = "prioriactions")
-sensitivity_data <- data.table::fread(sensitivity_path, data.table = FALSE)
+sensitivity_data <- data.table::fread(paste0(prioriactions_path,"sensibility.dat"), data.table = FALSE)
 head(sensitivity_data)
 
 ## Load in the boundary data
-bound_path <- system.file("extdata/input/bound.dat", package = "prioriactions")
-bound_data <- data.table::fread(bound_path, data.table = FALSE)
-head(bound_data)
+boundary_data <- data.table::fread(paste0(prioriactions_path,"bound.dat"), data.table = FALSE)
+head(boundary_data)
 
 ## Create data instance
 problem_data <- problem(
-  pu = pu_data, features = spec_data, rij = puvspr_data,
-  threats = threats_data, sensitivity = sensitivity_data,
-  bound = bound_data
+  pu = pu_data, features = features_data, dist_features = dist_features_data,
+  dist_threats = dist_threats_data, threats = threats_data, sensitivity = sensitivity_data,
+  boundary = boundary_data
 )
 
 ## Instance summary
